Move `FeedCachePolicy` To its own file
